+++ packages/ui-components/test-setup.ts
@@ -1,12 +1,8 @@
 import '@testing-library/jest-dom/vitest';
 import { vi } from 'vitest';
-import { mockRegistryStore } from './src/lib/__mocks__/stores';
+
 vi.mock('codemirror-rainlang', () => ({
 	RainlangLR: vi.fn()
-}));
-
-vi.mock('$lib/providers/registry/useRegistry', () => ({
-	useRegistry: vi.fn().mockReturnValue(mockRegistryStore)
 }));
 
 vi.mock('$app/stores', async () => {
